Crate mirrord_progress

Structs

FinishedTaskMessage 🔒

Message sent when a task is finished.

NewTaskMessage 🔒

Message sent when a new task is created using subtask/new

NoProgress

ProgressReport 🔒

Global progress state.

TaskProgress

Progress report for a single task.

Enums

ProgressMessage 🔒

ProgressMode

ProgressMode specifies the way progress is reported by TaskProgress.

Constants

MIRRORD_PROGRESS_ENV

The environment variable name that is used to determine the mode of progress reporting

Statics

PROGRESS 🔒

TASK_COUNT 🔒

Traits

Progress

Functions

get_report 🔒

init_from_env

Initialize progress reporting based on the MIRRORD_PROGRESS_ENV environment variable.